Try to open a new amqp channel when the old one is closed
Try to open a new amqp channel every time the old one is closed, except in
the case where the connection is closed.

The opening is make using the 'retry' pattern.

If is not possible to open a new amqp channel, we should close the 'messages'
channel, so that the user is not locked anymore.